v2.35: zero-copy evolution, primal constants, tarpc wiring, coverage push, docs cleanup
Zero-copy: bytes::Bytes for binary payloads with base64 serde helpers
Primal constants: centralized primal_names module, 15 files across 8 crates updated
tarpc transport: unix feature enabled, socket helpers in primal-sdk
Coverage: +183 tests (4,275 total), 0 failures, 0 clippy warnings
Large files: 6 test modules extracted for 1000-line compliance
Debt evolution: nix→rustix, sysinfo→/proc, JSON-RPC consolidated
Docs: updated root docs, CHANGELOG, EVOLUTION_ROADMAP to v2.35 metrics
Cleanup: archived start_nucleus.sh, fixed stale references, updated scripts README
ecoBin v3.0 compliant, AGPL-3.0-only, 0 unsafe code, 0 hardcoded primals

Made-with: Cursor

+++ crates/biomeos-api/tests/websocket_integration.rs
@@ -47,13 +47,14 @@ struct JsonRpcResponse {
     result: Option<serde_json::Value>,
     #[serde(default)]
     error: Option<JsonRpcError>,
+    #[serde(default)]
     id: Option<serde_json::Value>,
 }
 
 #[derive(Debug, serde::Deserialize)]
 #[allow(dead_code)]
 struct JsonRpcError {
-    code: i32,
+    code: i64,
     message: String,
 }
